Allow branch labels in node-data JSONs
Previously branch labels could not be specified in data passed to
`augur export v2` except for two "special cases":
(1) Mutations (stored in node-data-json -> nodes) would create branch
labels "aa", if applicable.
(2) `clade_annotation` (stored in node-data-json -> nodes) was
interpreted to be the "clade" branch label, and exported as such.

Here we extend the allowed node-data structure to include a top-level
key `branch_labels` as described in [1]. This data is exported
in the appropriate format for Auspice. This provides an way for
analyses to export custom branch labels other than the two special-
cases described above.

(Currently no augur commands can produce such data, but this will
change. See [1] for more.)

[1] https://github.com/nextstrain/augur/issues/720