Clade and branch-label names can be user-specified
Previously the `augur clades` command produced a node-data JSON
which stored clade membership as the node-trait "clade_membership"
and defined the basal nodes of each clade with the node-trait
"clade_annotation". `augur export v2` interpreted the latter as a
special-case and produced a branch label with the same name.

The previous commit allowed `augur export` to be supplied node-data
JSONs with a `branch_labels` structure. Here we update `augur clades`
to export data in this structure, which allows the user to specify
the keys to use.

To preserve backwards compatibility if neither key is specified, we
default to trait-name="clade_membership" and label-name="clade, which
will be exported from `augur export v2` correctly without needing
any configuration changes.

Closes https://github.com/nextstrain/augur/issues/720